requirements.txt
blinker==1.9.0

click==8.1.8

colorama==0.4.6

Flask==3.1.0

imbalanced-learn==0.13.0

itsdangerous==2.2.0

Jinja2==3.1.5

joblib==1.4.2

lightgbm==4.5.0

MarkupSafe==3.0.2

numpy==2.2.2

pandas==2.2.3

python-dateutil==2.9.0.post0

pytz==2024.2

scikit-learn==1.6.1

scipy==1.15.1

six==1.17.0

sklearn-compat==0.1.3

threadpoolctl==3.5.0

tzdata==2025.1

Werkzeug==3.1.3

python-box==6.0.2

pyYAML

tqdm

matplotlib

notebook

types-PyYAML

ensure

lightgbm

imbalanced-learn

-e .